Remove Banners component, have all invites in same location
