+++ packages/frontend-2/components/projects/DashboardHeader.vue
@@ -8,22 +8,27 @@
         v-if="projectsInvites?.projectInvites?.length"
         :invites="projectsInvites"
       />
-      <WorkspaceInviteBanners
-        v-if="
-          workspacesInvites?.workspaceInvites?.length ||
-          workspacesInvites?.discoverableWorkspaces?.length
-        "
-        :invites="workspacesInvites"
+      <WorkspaceInviteBanner
+        v-for="invite in workspaceInvites"
+        :key="invite.id"
+        :invite="invite"
+      />
+      <WorkspaceInviteDiscoverableWorkspaceBanner
+        v-for="workspace in discoverableWorkspaces"
+        :key="workspace.id"
+        :workspace="workspace"
       />
     </div>
   </div>
 </template>
 <script setup lang="ts">
+import { useSynchronizedCookie } from '~/lib/common/composables/reactiveCookie'
 import { graphql } from '~/lib/common/generated/gql'
 import type {
   ProjectsDashboardHeaderProjects_UserFragment,
   ProjectsDashboardHeaderWorkspaces_UserFragment
 } from '~/lib/common/generated/gql/graphql'
+import { CookieKeys } from '~/lib/common/helpers/constants'
 
 graphql(`
   fragment ProjectsDashboardHeaderProjects_User on User {
@@ -33,7 +38,12 @@ graphql(`
 
 graphql(`
   fragment ProjectsDashboardHeaderWorkspaces_User on User {
-    ...WorkspaceInviteBanners_User
+    discoverableWorkspaces {
+      ...WorkspaceInviteDiscoverableWorkspaceBanner_DiscoverableWorkspace
+    }
+    workspaceInvites {
+      ...WorkspaceInviteBanner_PendingWorkspaceCollaborator
+    }
   }
 `)
 
@@ -42,11 +52,26 @@ const props = defineProps<{
   workspacesInvites?: ProjectsDashboardHeaderWorkspaces_UserFragment
 }>()
 
+const dismissedDiscoverableWorkspaces = useSynchronizedCookie<string[]>(
+  CookieKeys.DismissedDiscoverableWorkspaces,
+  {
+    default: () => []
+  }
+)
+
+const workspaceInvites = computed(() => props.workspacesInvites?.workspaceInvites || [])
+const discoverableWorkspaces = computed(
+  () =>
+    props.workspacesInvites?.discoverableWorkspaces?.filter(
+      (workspace) => !dismissedDiscoverableWorkspaces.value.includes(workspace.id)
+    ) || []
+)
+
 const hasBanners = computed(() => {
   return (
     props.projectsInvites?.projectInvites?.length ||
-    props.workspacesInvites?.workspaceInvites?.length ||
-    props.workspacesInvites?.discoverableWorkspaces?.length
+    workspaceInvites.value.length ||
+    discoverableWorkspaces.value.length
   )
 })
 </script>
